feat: Copy relevant .env.stage file on build

+++ apps/node-server/scripts/copy-env.ts
@@ -1,9 +1,14 @@
 import { cpSync, existsSync } from 'fs';
+import { appendFileSync, readFileSync } from 'fs';
 import { resolve } from 'path';
 
-import { packageRootDir } from '../../../cdk/backend-server-cdk/src/location';
+import { packageRootDir } from '../src/location';
 
 const envFile = resolve(packageRootDir, '.env');
+const additionalEnvFile =
+  process.env.ENV === 'production'
+    ? resolve(packageRootDir, '.env.production')
+    : resolve(packageRootDir, '.env.dev');
 const distDirectory = resolve(packageRootDir, 'dist');
 
 if (!existsSync(distDirectory)) {
@@ -17,5 +22,16 @@ if (!existsSync(envFile)) {
   process.exit(0);
 }
 
+if (!existsSync(additionalEnvFile)) {
+  console.error(`❌ ${additionalEnvFile} file not found`);
+  // Do not cause build to fail
+  process.exit(0);
+}
+
 cpSync(envFile, `${distDirectory}/.env`);
+
+// Append to .env
+const additionalEnvContent = readFileSync(additionalEnvFile, 'utf-8');
+appendFileSync(`${distDirectory}/.env`, `\n${additionalEnvContent}`);
+
 console.info('✅ .env file copied to dist directory');
